Prevent various auto-boxing operations by using primitive getters.

+++ L2J_Mobius_4.0_GrandCrusade/java/org/l2jmobius/gameserver/enums/StatusUpdateType.java
@@ -70,6 +70,6 @@ public enum StatusUpdateType
 	
 	public int getValue(Creature creature)
 	{
-		return _valueSupplier.apply(creature);
+		return _valueSupplier.apply(creature).intValue();
 	}
 }
