+++ L2J_Mobius_CT_2.4_Epilogue/java/org/l2jmobius/gameserver/taskmanager/RandomAnimationTaskManager.java
@@ -46,7 +46,7 @@ public class RandomAnimationTaskManager
 			final long time = System.currentTimeMillis();
 			for (Entry<Npc, Long> entry : PENDING_ANIMATIONS.entrySet())
 			{
-				if (time > entry.getValue())
+				if (time > entry.getValue().longValue())
 				{
 					final Npc npc = entry.getKey();
 					if (npc.isInActiveRegion() && !npc.isDead() && !npc.isInCombat() && !npc.isMoving() && !npc.isStunned() && !npc.isSleeping() && !npc.isParalyzed())
